requirements.txt
google-genai
pydantic
python-dotenv
pillow
streamlit